chore: create a components directory and move files there
affects: @crave/farmblocks-more-info
