+++ packages/more-info/src/components/MoreInfo.test.js
