+++ packages/more-info/src/index.js
@@ -1,1 +1,1 @@
-export { default } from "./MoreInfo";
+export { default } from "./components/MoreInfo";
